+++ aimiaobi-20230801/alibabacloud_aimiaobi20230801/models.py
@@ -3640,6 +3640,139 @@ class DeleteMaterialByIdResponse(TeaModel):
         return self
 
 
+class DeleteStyleLearningResultRequest(TeaModel):
+    def __init__(
+        self,
+        agent_key: str = None,
+        id: int = None,
+    ):
+        # This parameter is required.
+        self.agent_key = agent_key
+        # This parameter is required.
+        self.id = id
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.agent_key is not None:
+            result['AgentKey'] = self.agent_key
+        if self.id is not None:
+            result['Id'] = self.id
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('AgentKey') is not None:
+            self.agent_key = m.get('AgentKey')
+        if m.get('Id') is not None:
+            self.id = m.get('Id')
+        return self
+
+
+class DeleteStyleLearningResultResponseBody(TeaModel):
+    def __init__(
+        self,
+        code: str = None,
+        data: bool = None,
+        http_status_code: int = None,
+        message: str = None,
+        request_id: str = None,
+        success: bool = None,
+    ):
+        self.code = code
+        self.data = data
+        self.http_status_code = http_status_code
+        self.message = message
+        self.request_id = request_id
+        self.success = success
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.code is not None:
+            result['Code'] = self.code
+        if self.data is not None:
+            result['Data'] = self.data
+        if self.http_status_code is not None:
+            result['HttpStatusCode'] = self.http_status_code
+        if self.message is not None:
+            result['Message'] = self.message
+        if self.request_id is not None:
+            result['RequestId'] = self.request_id
+        if self.success is not None:
+            result['Success'] = self.success
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Code') is not None:
+            self.code = m.get('Code')
+        if m.get('Data') is not None:
+            self.data = m.get('Data')
+        if m.get('HttpStatusCode') is not None:
+            self.http_status_code = m.get('HttpStatusCode')
+        if m.get('Message') is not None:
+            self.message = m.get('Message')
+        if m.get('RequestId') is not None:
+            self.request_id = m.get('RequestId')
+        if m.get('Success') is not None:
+            self.success = m.get('Success')
+        return self
+
+
+class DeleteStyleLearningResultResponse(TeaModel):
+    def __init__(
+        self,
+        headers: Dict[str, str] = None,
+        status_code: int = None,
+        body: DeleteStyleLearningResultResponseBody = None,
+    ):
+        self.headers = headers
+        self.status_code = status_code
+        self.body = body
+
+    def validate(self):
+        if self.body:
+            self.body.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.headers is not None:
+            result['headers'] = self.headers
+        if self.status_code is not None:
+            result['statusCode'] = self.status_code
+        if self.body is not None:
+            result['body'] = self.body.to_map()
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('headers') is not None:
+            self.headers = m.get('headers')
+        if m.get('statusCode') is not None:
+            self.status_code = m.get('statusCode')
+        if m.get('body') is not None:
+            temp_model = DeleteStyleLearningResultResponseBody()
+            self.body = temp_model.from_map(m['body'])
+        return self
+
+
 class DocumentExtractionRequest(TeaModel):
     def __init__(
         self,
@@ -26004,15 +26137,19 @@ class QueryAuditTaskResponseBodyData(TeaModel):
         self,
         audit_time: str = None,
         content: str = None,
+        html_content: str = None,
         response: QueryAuditTaskResponseBodyDataResponse = None,
         status: str = None,
         task_status: int = None,
+        title: str = None,
     ):
         self.audit_time = audit_time
         self.content = content
+        self.html_content = html_content
         self.response = response
         self.status = status
         self.task_status = task_status
+        self.title = title
 
     def validate(self):
         if self.response:
@@ -26028,12 +26165,16 @@ class QueryAuditTaskResponseBodyData(TeaModel):
             result['AuditTime'] = self.audit_time
         if self.content is not None:
             result['Content'] = self.content
+        if self.html_content is not None:
+            result['HtmlContent'] = self.html_content
         if self.response is not None:
             result['Response'] = self.response.to_map()
         if self.status is not None:
             result['Status'] = self.status
         if self.task_status is not None:
             result['TaskStatus'] = self.task_status
+        if self.title is not None:
+            result['Title'] = self.title
         return result
 
     def from_map(self, m: dict = None):
@@ -26042,6 +26183,8 @@ class QueryAuditTaskResponseBodyData(TeaModel):
             self.audit_time = m.get('AuditTime')
         if m.get('Content') is not None:
             self.content = m.get('Content')
+        if m.get('HtmlContent') is not None:
+            self.html_content = m.get('HtmlContent')
         if m.get('Response') is not None:
             temp_model = QueryAuditTaskResponseBodyDataResponse()
             self.response = temp_model.from_map(m['Response'])
@@ -26049,6 +26192,8 @@ class QueryAuditTaskResponseBodyData(TeaModel):
             self.status = m.get('Status')
         if m.get('TaskStatus') is not None:
             self.task_status = m.get('TaskStatus')
+        if m.get('Title') is not None:
+            self.title = m.get('Title')
         return self